+++ drivers/docker/config_test.go
@@ -21,10 +21,11 @@ func TestConfig_ParseHCL(t *testing.T) {
 				image = "redis:3.2"
 			}`,
 			&TaskConfig{
-				Image:        "redis:3.2",
-				Devices:      []DockerDevice{},
-				Mounts:       []DockerMount{},
-				CPUCFSPeriod: 100000,
+				Image:            "redis:3.2",
+				Devices:          []DockerDevice{},
+				Mounts:           []DockerMount{},
+				CPUCFSPeriod:     100000,
+				ImagePullTimeout: "5m",
 			},
 		},
 	}
@@ -53,40 +54,44 @@ func TestConfig_ParseJSON(t *testing.T) {
 			name:  "nil values for blocks are safe",
 			input: `{"Config": {"image": "bash:3", "mounts": null}}`,
 			expected: TaskConfig{
-				Image:        "bash:3",
-				Mounts:       []DockerMount{},
-				Devices:      []DockerDevice{},
-				CPUCFSPeriod: 100000,
+				Image:            "bash:3",
+				Mounts:           []DockerMount{},
+				Devices:          []DockerDevice{},
+				CPUCFSPeriod:     100000,
+				ImagePullTimeout: "5m",
 			},
 		},
 		{
 			name:  "nil values for 'volumes' field are safe",
 			input: `{"Config": {"image": "bash:3", "volumes": null}}`,
 			expected: TaskConfig{
-				Image:        "bash:3",
-				Mounts:       []DockerMount{},
-				Devices:      []DockerDevice{},
-				CPUCFSPeriod: 100000,
+				Image:            "bash:3",
+				Mounts:           []DockerMount{},
+				Devices:          []DockerDevice{},
+				CPUCFSPeriod:     100000,
+				ImagePullTimeout: "5m",
 			},
 		},
 		{
 			name:  "nil values for 'args' field are safe",
 			input: `{"Config": {"image": "bash:3", "args": null}}`,
 			expected: TaskConfig{
-				Image:        "bash:3",
-				Mounts:       []DockerMount{},
-				Devices:      []DockerDevice{},
-				CPUCFSPeriod: 100000,
+				Image:            "bash:3",
+				Mounts:           []DockerMount{},
+				Devices:          []DockerDevice{},
+				CPUCFSPeriod:     100000,
+				ImagePullTimeout: "5m",
 			},
 		},
 		{
 			name:  "nil values for string fields are safe",
 			input: `{"Config": {"image": "bash:3", "command": null}}`,
 			expected: TaskConfig{
-				Image:        "bash:3",
-				Mounts:       []DockerMount{},
-				Devices:      []DockerDevice{},
-				CPUCFSPeriod: 100000,
+				Image:            "bash:3",
+				Mounts:           []DockerMount{},
+				Devices:          []DockerDevice{},
+				CPUCFSPeriod:     100000,
+				ImagePullTimeout: "5m",
 			},
 		},
 	}
@@ -178,6 +183,7 @@ func TestConfig_ParseAllHCL(t *testing.T) {
 	cfgStr := `
 config {
   image = "redis:3.2"
+  image_pull_timeout = "15m"
   advertise_ipv6_address = true
   args = ["command_arg1", "command_arg2"]
   auth {
@@ -302,6 +308,7 @@ config {
 
 	expected := &TaskConfig{
 		Image:             "redis:3.2",
+		ImagePullTimeout:  "15m",
 		AdvertiseIPv6Addr: true,
 		Args:              []string{"command_arg1", "command_arg2"},
 		Auth: DockerAuth{
@@ -526,6 +533,33 @@ func TestConfig_InternalCapabilities(t *testing.T) {
 
 			d := &Driver{config: &tc}
 			require.Equal(t, c.expected, d.InternalCapabilities())
+		})
+	}
+}
+
+func TestConfig_DriverConfig_InfraImagePullTimeout(t *testing.T) {
+	cases := []struct {
+		name     string
+		config   string
+		expected string
+	}{
+		{
+			name:     "default",
+			config:   `{}`,
+			expected: "5m",
+		},
+		{
+			name:     "set explicitly",
+			config:   `{ infra_image_pull_timeout = "1m" }`,
+			expected: "1m",
+		},
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(t *testing.T) {
+			var tc DriverConfig
+			hclutils.NewConfigParser(configSpec).ParseHCL(t, "config "+c.config, &tc)
+			require.Equal(t, c.expected, tc.InfraImagePullTimeout)
 		})
 	}
 }