+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.12.4 (Unreleased)
+
+IMPROVEMENTS:
+
+ * api: Added node purge SDK functionality. [[GH-8142](https://github.com/hashicorp/nomad/issues/8142)]
+ * driver/docker: Allow configurable image pull context timeout setting. [[GH-5718](https://github.com/hashicorp/nomad/issues/5718)]
+
 ## 0.12.3 (August 13, 2020)
 
 BUG FIXES: